* commit '4fcfa14ea5f6496de1e17cc1445ef094fea3fc3d': (79 commits)
  omit unused code
  change buildStatus success and failure strings
  ensure that jenkins will email after pipeline is fixed
  using ssh instead of http for git url
  remove validation line
  validation
  delete dist folder and ignore hidden branch files
  Clean before preparing a version number
  update version number for jenkins testing
  remove path editing for jenkins-scripts
  move declarations out of try block
  change jenkins-scripts path
  put ant tool into path only in stages where it is necessary
  change path and echo for debugging
  add jenkins scripts to path
  Add email, slack, and version tagging
  add parameterization to jenkins builds
  make python_executable a string property, not path
  debugging
  faulty dependencies
  ...

+++ test/testOmeTifWriter.py
@@ -18,7 +18,7 @@ class OmeTifWriterTestGroup(unittest.TestCase):
         imageio.init()
         cls.dir_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'img')
         cls.file = os.path.join(cls.dir_path, 'ometif_test_output.ome.tif')
-        cls.image = np.random.rand(1, 40, 3, 128, 256)
+        cls.image = np.random.rand(1, 40, 3, 128, 256).astype('int')
         cls.writer = omeTifWriter.OmeTifWriter(cls.file, overwrite_file=True)
         if not os.path.isfile(cls.file):
             open(cls.file, 'a').close()
